リモートリポジトリの変更

+++ src/database/migrations/2024_11_23_143706_create_items_table.php
@@ -19,7 +19,7 @@ class CreateItemsTable extends Migration
             $table->foreignId('condition_id')->constrained()->cascadeOnDelete();
             $table->string('product_name', 255)->comment('商品名');
             $table->string('brand_name', 255)->comment('ブランド名');
-            $table->integer('price')->comment('商品の価格');
+            $table->bigInteger('price')->comment('商品の価格');
             $table->text('description')->comment('商品の説明');
             $table->string('img_url')->comment('商品画像');
             $table->timestamps();
